+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Bugfree.Spo.ExternalSharingCenter 1.0.1 (2016-07-31)
+
+* [#3](https://github.com/ronnieholm/Bugfree.Spo.ExternalSharingCenter/issues/3) - External sharings with InvitedAs != AcceptedAs causes warning mail generation to fail
+
 ## Bugfree.Spo.ExternalSharingCenter 1.0.0 (2016-05-09)
 
 * Initial release